Mudanças organizacionais na estrutura de arquivos e controllers
