+++ src/main/java/com/projetoOO/projetoOO/projController/IndexController.java
@@ -1,0 +1,16 @@
+package com.projetoOO.projetoOO.projController;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+
+@Controller
+public class IndexController {
+	@RequestMapping("/")
+	public String index() {
+		return "index";
+	}	
+	
+}
